feat: suivre_renvoi cross-reference-following tool (7b)
Second MCP tool: given an article, returns the references it makes to
other articles, split into resolved-and-currently-visible (renvois) vs.
not followable right now (nonResolus), enforcing the same "a repealed
article must never be returned" rule chercher_droit's RLS already applies -
here per-target rather than per-query, via a LEFT JOIN under the same
anon/RLS session SupabaseRetriever uses.

Verified end-to-end against real Supabase data, including a live proof
that the per-target visibility filter genuinely works (an ancient
dateReference moves an otherwise-resolved renvoi into nonResolus).

Two audit findings repaired and re-reviewed before completing: F-01
(fetchRenvoiRows exported without enforcing its RLS-session precondition -
renamed to make misuse obvious) and F-02 (Chunk.id silently returned as a
string instead of a number from SupabaseRetriever, the same bigint-cast
bug this feature hit and fixed for renvois.id, found in already-shipped
packages/retrieval code while reviewing this feature's own fix).

Archives the spec to blueprint/history/features/07b-suivre-renvoi.md and
checks off 7b in build-plan.md. Item 7 (Tool server) stays unchecked -
7c-7d remain.

+++ packages/retrieval/src/supabase-retriever.ts
@@ -9,6 +9,10 @@ const PRE_FUSION_LIMIT = 50; // top 50 de chaque liste avant fusion, indépendan
 // vector_search et keyword_search interrogent chunks directement, donc la
 // RLS de 4c (chunks_search_read) s'applique déjà dans les deux CTE - le
 // join final vers chunks est redondant mais sans risque.
+// c.id::int (F-02, trouvé en 7b) : chunks.id est un bigint (identity) - pg
+// le renverrait sinon en chaîne (précision au-delà de Number.MAX_SAFE_
+// INTEGER), ce qui violerait silencieusement Chunk.id: number. Un id de
+// chunk ne s'approchera jamais de la limite d'un int4.
 const HYBRID_SEARCH_SQL = `
   with vector_search as (
     select id, row_number() over (order by embedding <=> $1::extensions.vector) as rank
@@ -30,7 +34,7 @@ const HYBRID_SEARCH_SQL = `
     from vector_search v
     full outer join keyword_search k on v.id = k.id
   )
-  select c.id, c.article_identifier, c.subdivision_label, c.contenu, fused.score
+  select c.id::int, c.article_identifier, c.subdivision_label, c.contenu, fused.score
   from fused join chunks c on c.id = fused.id
   order by fused.score desc
   limit $3
